First iteration on the unified loading bar
