+++ client/landing/stepper/declarative-flow/internals/components/stepper-loader/style.scss
@@ -1,3 +1,32 @@
+@import "@wordpress/base-styles/breakpoints";
+@import "@wordpress/base-styles/mixins";
+@import "@automattic/onboarding/styles/mixins";
+
 .stepper-loader {
-	animation: loading-fade 1.6s ease-in-out infinite;
+	--wp-components-color-foreground: #3858e9;
+	padding: 1em;
+	max-width: 540px;
+	text-align: center;
+	margin: 32vh auto 0;
+
+	.stepper-loader__title {
+		@include onboarding-font-recoleta;
+		font-weight: 400;
+		height: 40px;
+		letter-spacing: -0.4px;
+		line-height: 40px;
+		margin: 0;
+		text-align: center;
+		vertical-align: middle;
+
+		font-size: $font-title-medium;
+
+		@include break-medium {
+			font-size: $font-title-large;
+		}
+	}
+
+	.stepper-loader__progress-bar {
+		margin: 46px auto 0 auto;
+	}
 }
