+++ client/landing/stepper/declarative-flow/internals/index.tsx
@@ -6,6 +6,7 @@ import Modal from 'react-modal';
 import { generatePath, useParams } from 'react-router';
 import { Route, Routes } from 'react-router-dom';
 import DocumentHead from 'calypso/components/data/document-head';
+import WordPressLogo from 'calypso/components/wordpress-logo';
 import { STEPPER_INTERNAL_STORE } from 'calypso/landing/stepper/stores';
 import AsyncCheckoutModal from 'calypso/my-sites/checkout/modal/async';
 import { useSelector } from 'calypso/state';
@@ -156,7 +157,10 @@ export const FlowRenderer: React.FC< { flow: Flow; steps: readonly StepperStep[]
 	const renderStep = ( step: StepperStep ) => {
 		switch ( assertCondition.state ) {
 			case AssertConditionState.CHECKING:
-				return <StepperLoader />;
+				if ( isWooExpressFlow( flow.name ) ) {
+					return <WordPressLogo size={ 72 } className="wpcom-site__logo stepper-loader" />;
+				}
+				return <StepperLoader flow={ flow.name } />;
 			case AssertConditionState.FAILURE:
 				return null;
 		}
