+++ client/landing/stepper/declarative-flow/internals/steps-repository/processing-step/index.tsx
@@ -5,19 +5,16 @@ import {
 	isNewSiteMigrationFlow,
 	isUpdateDesignFlow,
 	ECOMMERCE_FLOW,
-	isWooExpressFlow,
-	isTransferringHostedSiteCreationFlow,
 	HUNDRED_YEAR_DOMAIN_FLOW,
 	HUNDRED_YEAR_PLAN_FLOW,
 	HUNDRED_YEAR_DOMAIN_TRANSFER,
 	isAnyHostingFlow,
 } from '@automattic/onboarding';
-import { ProgressBar } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
 import { useI18n } from '@wordpress/react-i18n';
 import { useEffect, useState, useRef } from 'react';
 import DocumentHead from 'calypso/components/data/document-head';
-import { LoadingBar } from 'calypso/components/loading-bar';
+import { StepperLoader } from 'calypso/landing/stepper/declarative-flow/internals/components';
 import availableFlows from 'calypso/landing/stepper/declarative-flow/registered-flows';
 import { useRecordSignupComplete } from 'calypso/landing/stepper/hooks/use-record-signup-complete';
 import { ONBOARD_STORE } from 'calypso/landing/stepper/stores';
@@ -33,6 +30,7 @@ import TailoredFlowPreCheckoutScreen from './tailored-flow-precheckout-screen';
 import type { StepProps } from '../../types';
 import type { OnboardSelect } from '@automattic/data-stores';
 import './style.scss';
+
 interface ProcessingStepProps extends StepProps {
 	title?: string;
 	subtitle?: string;
@@ -195,26 +193,6 @@ const ProcessingStep: React.FC< ProcessingStepProps > = function ( props ) {
 		return <HundredYearPlanFlowProcessingScreen />;
 	}
 
-	const subtitle = getSubtitle();
-
-	const renderProgressComponent = () => {
-		if ( isWooExpressFlow( flow ) || isTransferringHostedSiteCreationFlow( flow ) ) {
-			return (
-				<LoadingBar
-					progress={ progress }
-					className="processing-step__content woocommerce-install__content"
-				/>
-			);
-		}
-
-		return (
-			<ProgressBar
-				value={ progress >= 0 ? progress * 100 : undefined }
-				className="processing-step__progress-bar"
-			/>
-		);
-	};
-
 	return (
 		<>
 			<DocumentHead title={ __( 'Processing' ) } />
@@ -223,13 +201,12 @@ const ProcessingStep: React.FC< ProcessingStepProps > = function ( props ) {
 				hideFormattedHeader
 				stepName="processing-step"
 				stepContent={
-					<>
-						<div className="processing-step">
-							<h1 className="processing-step__progress-step">{ getCurrentMessage() }</h1>
-							{ renderProgressComponent() }
-							{ subtitle && <p className="processing-step__subtitle">{ subtitle }</p> }
-						</div>
-					</>
+					<StepperLoader
+						title={ getCurrentMessage() }
+						subtitle={ getSubtitle() }
+						progress={ progress }
+						flow={ flow }
+					/>
 				}
 				recordTracksEvent={ recordTracksEvent }
 				showJetpackPowered={ isJetpackPowered }
